1、将保存在redis中的浏览历史数据类型由list改成zset 2、完善goods模块和order模块 3、user模块新增统计广告和通知的浏览次数 4、修改日志文件

+++ irh-service/irh-service-goods/src/main/java/top/imuster/goods/service/ProductRecordService.java
@@ -41,5 +41,5 @@ public interface ProductRecordService {
      * @param index
      * @reture: top.imuster.common.base.wrapper.Message<java.lang.String>
      **/
-    Message<String> deleteByIndex(Integer index, Long userId);
+    Message<String> deleteByIndex(Long index, Long userId);
 }
